requirements.txt
requests==2.31.0
PyPDF2==3.0.1
Pillow==10.0.0 